Fix and move client query and commands
