+++ src/Feedback.Analyzer.Backend/Feedback.Analyzer.Application/Clients/Queries/ClientGetByIdQuery.cs
@@ -1,12 +1,13 @@
 ﻿using Feedback.Analyzer.Domain.Common.Commands;
+using Feedback.Analyzer.Domain.Common.Queries;
 using Feedback.Analyzer.Domain.Entities;
 
-namespace Feedback.Analyzer.Application.Clients.Commands;
+namespace Feedback.Analyzer.Application.Clients.Queries;
 
 /// <summary>
 /// Represents a command to retrieve a client by their unique identifier.
 /// </summary>
-public class ClientGetByIdCommand : ICommand<Client?>
+public class ClientGetByIdQuery : IQuery<Client>
 {
     public Guid ClientId { get; set; }
 }
