+++ src/Feedback.Analyzer.Backend/Feedback.Analyzer.Domain/Common/Queries/IQueryHandler.cs
@@ -7,6 +7,4 @@ namespace Feedback.Analyzer.Domain.Common.Queries;
 /// </summary>
 /// <typeparam name="TQuery">The type of query being handled.</typeparam>
 /// <typeparam name="TResult">The type of result produced by handling the query.</typeparam>
-public interface IQueryHandler<in TQuery, TResult> : IRequestHandler<TQuery, TResult> where TQuery : IQuery<TResult>
-{
-}
+public interface IQueryHandler<in TQuery, TResult> : IRequestHandler<TQuery, TResult> where TQuery : IQuery<TResult>;
