--- src/Feedback.Analyzer.Backend/Feedback.Analyzer.Infrastructure/CommandHandlers/ClientGetByIdCommandHandler.cs
@@ -1,18 +1,0 @@
-﻿using Feedback.Analyzer.Application.Clients.Commands;
-using Feedback.Analyzer.Application.Clients.Services;
-using Feedback.Analyzer.Domain.Common.Commands;
-using Feedback.Analyzer.Domain.Common.Queries;
-using Feedback.Analyzer.Domain.Entities;
-
-namespace Feedback.Analyzer.Infrastructure.CommandHandlers;
-
-/// <summary>
-/// Command handler for retrieving a client by their unique identifier.
-/// </summary>
-public class ClientGetByIdCommandHandler(IClientService clientService) : ICommandHandler<ClientGetByIdCommand, Client?>
-{
-    public async Task<Client?> Handle(ClientGetByIdCommand request, CancellationToken cancellationToken)
-    {
-        return await clientService.GetByIdAsync(request.ClientId, new QueryOptions() { AsNoTracking = true }, cancellationToken);
-    }
-}
